Moves headers to separate directory
