+++ src/CMakeLists.txt
@@ -7,18 +7,9 @@ target_link_libraries(jsonish_test PRIVATE taocpp::pegtl)
 
 add_library(config_reader
   STATIC
-  config_actions.h
-  config_classes.h
-  config_exceptions.h
-  config_grammar.h
-  config_helpers.h
-  config_reader.h
-  config_selector.h
+  ${CFG_HEADERS}
   config_helpers.cpp
   config_reader.cpp
-  math_actions.h
-  math_grammar.h
-  math_helpers.h
   math_helpers.cpp
 )
 
@@ -30,7 +21,8 @@ target_link_libraries(config_reader
 )
 
 target_include_directories(config_reader PUBLIC
-  ${PROJECT_SOURCE_DIR}/cpp/
+  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+  $<INSTALL_INTERFACE:include>
 )
 
 add_clang_format(config_reader)
@@ -44,6 +36,9 @@ target_link_libraries(config_build
   PRIVATE
   config_reader
 )
+target_include_directories(config_build PRIVATE
+  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+)
 add_clang_format(config_build)
 
 add_executable(config_reader_example config_reader_example.cpp)
@@ -52,10 +47,16 @@ target_link_libraries(config_reader_example
   config_reader
   fmt::fmt
 )
+target_include_directories(config_reader_example PRIVATE
+  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+)
 add_clang_format(config_reader_example)
 
 add_executable(include_file_test include_file_test.cpp)
 target_link_libraries(include_file_test PRIVATE taocpp::pegtl)
+target_include_directories(include_file_test PRIVATE
+  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+)
 
 
 install(TARGETS jsonish_test config_build config_reader_example include_file_test
