+++ src/config_build.cpp
@@ -6,8 +6,8 @@
 #include <magic_enum.hpp>
 #include <span>
 
-#include "config_reader.h"
-#include "logger.h"
+#include "cfg_reader/config_reader.h"
+#include "cfg_reader/logger.h"
 
 auto main(int argc, char* argv[]) -> int {
   try {
