+++ src/config_helpers.cpp
@@ -1,4 +1,4 @@
-#include "config_helpers.h"
+#include "cfg_reader/config_helpers.h"
 
 #include <fmt/format.h>
 
@@ -13,12 +13,12 @@
 #include <regex>
 #include <span>
 
-#include "config_actions.h"
-#include "config_classes.h"
-#include "config_exceptions.h"
-#include "logger.h"
-#include "math_actions.h"
-#include "utils.h"
+#include "cfg_reader/config_actions.h"
+#include "cfg_reader/config_classes.h"
+#include "cfg_reader/config_exceptions.h"
+#include "cfg_reader/logger.h"
+#include "cfg_reader/math_actions.h"
+#include "cfg_reader/utils.h"
 
 namespace {
 constexpr bool CONFIG_HELPERS_DEBUG{false};
