+++ src/include_file_test.cpp
@@ -10,7 +10,7 @@
 #include <tao/pegtl/contrib/trace.hpp>
 #include <vector>
 
-#include "config_grammar.h"
+#include "cfg_reader/config_grammar.h"
 
 namespace peg = TAO_PEGTL_NAMESPACE;
