+++ tests/CMakeLists.txt
@@ -13,7 +13,7 @@ target_link_libraries(
   )
 
 target_include_directories(utils_test PRIVATE
-  ${CMAKE_SOURCE_DIR}/cpp/
+  ${CMAKE_SOURCE_DIR}/include/
   )
 
 add_clang_format(utils_test)
@@ -35,7 +35,7 @@ target_link_libraries(
   )
 
 target_include_directories(config_helpers_test PRIVATE
-  ${CMAKE_SOURCE_DIR}/cpp/
+  ${CMAKE_SOURCE_DIR}/include/
   )
 
 add_clang_format(config_helpers_test)
@@ -58,7 +58,7 @@ target_link_libraries(
   )
 
 target_include_directories(config_grammar_test PRIVATE
-  ${CMAKE_SOURCE_DIR}/cpp/
+  ${CMAKE_SOURCE_DIR}/include/
   )
 
 add_clang_format(config_grammar_test)
@@ -73,15 +73,12 @@ add_executable(
 target_link_libraries(
   config_exception_test
   config_reader
-  #fmt::fmt
-  #magic_enum::magic_enum
-  #range-v3::range-v3
   taocpp::pegtl
   gtest_main
   )
 
 target_include_directories(config_exception_test PRIVATE
-  ${CMAKE_SOURCE_DIR}/cpp/
+  ${CMAKE_SOURCE_DIR}/include/
   )
 
 add_clang_format(config_exception_test)
@@ -101,7 +98,7 @@ target_link_libraries(
   )
 
 target_include_directories(config_parse_test PRIVATE
-  ${CMAKE_SOURCE_DIR}/cpp/
+  ${CMAKE_SOURCE_DIR}/include/
   )
 
 add_clang_format(config_parse_test)
@@ -123,7 +120,7 @@ target_link_libraries(
   )
 
 target_include_directories(math_test PRIVATE
-  ${CMAKE_SOURCE_DIR}/cpp/
+  ${CMAKE_SOURCE_DIR}/include/
   )
 
 add_clang_format(math_test)
