+++ tests/config_exception_test.cpp
@@ -7,11 +7,11 @@
 #include <tao/pegtl.hpp>
 #include <tao/pegtl/contrib/parse_tree.hpp>
 
-#include "config_actions.h"
-#include "config_exceptions.h"
-#include "config_grammar.h"
-#include "config_reader.h"
-#include "logger.h"
+#include "cfg_reader/config_actions.h"
+#include "cfg_reader/config_exceptions.h"
+#include "cfg_reader/config_grammar.h"
+#include "cfg_reader/config_reader.h"
+#include "cfg_reader/logger.h"
 
 namespace {
