+++ tests/utils_test.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "cfg_reader/utils.h"
 
 #include <gtest/gtest.h>
